requirements.txt
# Core dependencies for Local AI Packaged
requests==2.32.5
python-dotenv==1.0.0
pytest==8.3.3
pytest-cov==5.0.0
pyyaml==6.0.2
asyncio-mqtt==0.16.0
neo4j==5.24.0
psycopg2-binary==2.9.9
qdrant-client==1.11.0
langchain==0.3.3
langchain-community==0.3.2
langchain-openai==0.2.1
openai==1.51.2
